+++ src/Components/Builder/Builder.vue
@@ -4,7 +4,14 @@
     <AppletSchemaBuilder
       v-else
       exportButton
+      :initialData="(isEditing || null) && currentApplet"
+      :key="componentKey"
+      :getProtocols="getProtocols"
+      :versions="versions"
       @uploadProtocol="onUploadProtocol"
+      @updateProtocol="onUpdateProtocol"
+      @prepareApplet="onPrepareApplet"
+      @onUploadError="onUploadError"
     />
 
     <Information
@@ -49,7 +56,29 @@ export default {
       dialogTitle: "",
       appletPasswordDialog: false,
       newApplet: {},
+      isEditing: false,
+      versions: [],
+      componentKey: 1
     };
+  },
+  async beforeMount() {
+    const apiHost = this.$store.state.backend;
+    const token = this.$store.state.auth.authToken.token;
+    const appletId = this.currentApplet.applet._id.split('/')[1];
+
+    this.versions = [];
+
+    if (this.$route.params.isEditing) {
+      this.isEditing = true;
+
+      const resp = await api.getAppletVersions({ apiHost, token, appletId });
+      this.versions = resp.data;
+    }
+  },
+  computed: {
+    currentApplet() {
+      return this.$store.state.currentApplet;
+    }
   },
   methods: {
     onClickSubmitPassword(appletPassword) {
@@ -62,7 +91,7 @@ export default {
     },
     addNewApplet(appletPassword) {
       const protocol = new FormData();
-      protocol.set("protocol", JSON.stringify(newApplet || {}));
+      protocol.set("protocol", JSON.stringify(this.newApplet || {}));
 
       api
         .createApplet({
@@ -72,17 +101,77 @@ export default {
           apiHost: this.$store.state.backend,
         })
         .then(resp => {
-          this.dialogText = resp.data.message;
-          this.dialogTitle = 'Upload Received';
-          this.dialog = true;
+          this.onUploadSucess(resp.data.message);
         })
         .catch((e) => {
           console.log(e);
-          this.dialogText = "There was an error uploading your applet. Please try again or report the issue.";
-          this.dialogTitle = 'Upload Error';
-          this.dialog = true;
+          this.onUploadError();
         });
     },
-  },
+    onUpdateProtocol(updateData) {
+      const protocol = new FormData();
+      protocol.set("protocol", JSON.stringify(updateData|| {}));
+
+      const appletId = this.currentApplet.applet._id.split('/')[1];
+      const token = this.$store.state.auth.authToken.token;
+      const apiHost = this.$store.state.backend;
+      api.updateApplet({
+        data: protocol,
+        appletId,
+        token,
+        apiHost,
+      }).then(resp => {
+        this.$store.commit('updateAppletData', {
+          ...resp.data,
+          roles: this.currentApplet.roles
+        });
+
+        api.getAppletVersions({ apiHost, token, appletId }).then(resp => {
+          this.versions = resp.data;
+          this.componentKey = this.componentKey + 1;
+        });
+
+        this.onUploadSucess();
+      }).catch(e => {
+        this.onUploadError();
+      })
+    },
+    onPrepareApplet(data) {
+      const protocol = new FormData();
+      protocol.set("protocol", JSON.stringify(data || {}));
+
+      const appletId = this.currentApplet.applet._id.split('/')[1];
+      const token = this.$store.state.auth.authToken.token;
+      const apiHost = this.$store.state.backend;
+
+      api.prepareApplet({
+        apiHost, token, data: protocol, appletId
+      }).then(resp => api.getAppletVersions({ apiHost, token, appletId })).then(resp => {
+        this.versions = resp.data;
+        this.componentKey = this.componentKey + 1;
+      });
+    },
+    onUploadSucess() {
+        this.dialogText = 'Your applet is successfully updated';
+        this.dialogTitle = 'Upload Received';
+        this.dialog = true;
+    },
+    onUploadError(msg) {
+        this.dialogText = msg || "There was an error uploading your applet. Please try again or report the issue.";
+        this.dialogTitle = 'Upload Error';
+        this.dialog = true;
+    },
+    getProtocols(versions) {
+      if (!this.isEditing) return [];
+      const appletId = this.currentApplet.applet._id.split('/')[1];
+
+      return api.getProtocolData({
+        apiHost: this.$store.state.backend,
+        token: this.$store.state.auth.authToken.token,
+        appletId,
+        versions
+      })
+    }
+  }
 };
 </script>