+++ src/Components/Utils/api/api.vue
@@ -254,6 +254,26 @@ const createApplet = ({ apiHost, token, email, data }) =>
     data,
   });
 
+const updateApplet = ({ apiHost, token, data, appletId }) =>
+  axios({
+    method: "PUT",
+    url: `${apiHost}/applet/${appletId}/fromJSON`,
+    headers: {
+      "Girder-Token": token,
+    },
+    data,
+  })
+
+const prepareApplet = ({ apiHost, token, data, appletId }) =>
+  axios({
+    method: "PUT",
+    url: `${apiHost}/applet/${appletId}/prepare`,
+    headers: {
+      "Girder-Token": token,
+    },
+    data,
+  })
+
 const updateRegistration = ({ apiHost, token, groupId, open }) =>
   axios({
     method: "PUT",
@@ -319,6 +339,23 @@ const setAppletEncryption = ({ apiHost, token, appletId, data }) => axios({
     'Girder-Token': token,
   },
   data
+});
+
+const getAppletVersions = ({ apiHost, token, appletId }) => axios({
+  method: 'get',
+  url: `${apiHost}/applet/${appletId}/versions`,
+  headers: {
+    'Girder-Token': token
+  }
+});
+
+const getProtocolData = ({ apiHost, token, appletId, versions }) => axios({
+  method: 'get',
+  url: `${apiHost}/applet/${appletId}/protocolData`,
+  headers: {
+    'Girder-Token': token
+  },
+  params: { versions: JSON.stringify(versions) },
 })
 
 export default {
@@ -352,5 +389,9 @@ export default {
   duplicateApplet,
   replaceResponseData,
   setAppletEncryption,
+  updateApplet,
+  getAppletVersions,
+  getProtocolData,
+  prepareApplet,
 }
 </script>