+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.18] = 2020-09-28
+## [0.3.19] = 2020-09-30
 ### Changed
 - Fixed duplicated definition of userData
+
+## [0.3.18] = 2020-09-30
+### Changed
+- insert edit applet functionality
+- fix issues in response encryption
 
 ## [0.3.17] = 2020-09-22
 ### Changed